% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_glints.R
\name{desc_glints}
\alias{desc_glints}
\title{Description formatting for Glints}
\usage{
desc_glints(url, head, title, mode = "print")
}
\arguments{
\item{url}{Character of job's URLs.}

\item{head}{Include title. Default ot FALSE.}

\item{title}{Character title, same length with url.}

\item{mode}{The way to return the description text c("print", "telegram").}
}
\description{
Scrape Glints' job description (experimental).
}
\examples{
\dontrun{
desc_glints("https://glints.com/id/opportunities/jobs/0850c687-5e79-4402-ad84-d5c04cb422d6")
}

}
