% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fields_available}
\alias{fields_available}
\title{Available fields for each dataset}
\format{
A data frame with \code{data.frame} format
}
\usage{
fields_available
}
\description{
Comparison table describing available fields between dataset created.
}
\keyword{datasets}
