% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glints.R
\name{glints}
\alias{glints}
\title{Glints Vacancy}
\usage{
glints(key, limit = 30L)
}
\arguments{
\item{key}{Characters indicate the keyword for job search}

\item{limit}{Number of limit for job results}
}
\description{
Get job vacancy from Glints' website.
}
\examples{
\dontrun{
glints("data analyst", 15) # return data analyst job
}

}
