% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glints_desc.R
\name{glints_desc}
\alias{glints_desc}
\title{Description formatting for Glints}
\usage{
glints_desc(url)
}
\arguments{
\item{url}{(character) Job URL(s).}
}
\description{
Using descriptionRaw column as provided by API or alternatively
   using get_joburl function for scraping.
}
