% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glintsjobid}
\alias{glintsjobid}
\title{Glints job category ID}
\format{
A data frame with \code{tibble} format in two fields:
\describe{
  \item{id}{The id for job category}
  \item{name}{The name of category}
}
}
\source{
\url{https://glints.com/}
}
\usage{
glintsjobid
}
\description{
A dataset containing the id of Glints' job category and its name.
}
\keyword{datasets}
