% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gql.R
\name{gql}
\alias{gql}
\title{GraphQL Query Request}
\source{
\href{https://gist.github.com/rentrop/83cb1d8fc8593726a808032e55314019/}{Rentrop's Gist}
}
\usage{
gql(query, ..., token = NULL, var = NULL, opnam = NULL, url = url)
}
\arguments{
\item{query}{GraphQL query}

\item{...}{Other arguments used by \code{httr::POST()}}

\item{token}{Token}

\item{var}{Variables}

\item{opnam}{Operation name}

\item{url}{Complete URL to RestAPI service base to endpoint}
}
\value{
A list of query result
}
\description{
Request data using GraphQL query
}
