% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indeed.R
\name{indeed}
\alias{indeed}
\title{Indeed Indonesia Vacancy}
\usage{
indeed(key, limit = 30L)
}
\arguments{
\item{key}{(character) Keyword for the jobs}

\item{limit}{(numeric) Limit amount of job results}
}
\value{
Job vacancy data.frame in tibble format
}
\description{
Get job vacancy from Indeed's website (Indonesia)
}
\examples{
\dontrun{
indeed("ux designer") # return UX designer job
}

}
