% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobdescform.R
\name{jobdescform}
\alias{jobdescform}
\title{Job Description Formation}
\usage{
jobdescform(urls, format = "md")
}
\arguments{
\item{urls}{vector of url list}
}
\value{
vector text explanation contained jobdesc
}
\description{
Job Description Formation
}
