% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobstreet.R
\name{jobstreet}
\alias{jobstreet}
\title{Get Vacancy from Jobstreet}
\usage{
jobstreet(key = "data analyst", page = 1)
}
\arguments{
\item{key}{Keyword for searching}

\item{page}{Pages to be looking at}
}
\value{
Job opportunity data.frame in tibble format
}
\description{
Get Vacancy from Jobstreet
}
\examples{
\dontrun{
jobstreet("data analyst", 1)
}
}
