% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{set_default_key}
\alias{set_default_key}
\title{Set default keyword message}
\usage{
set_default_key(key)
}
\arguments{
\item{key}{Characters indicate the keyword for job search}
}
\description{
Supply key argument
}
