% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glints.R
\name{glints}
\alias{glints}
\title{Glints Vacancy}
\usage{
glints(jobid, limit = 30)
}
\arguments{
\item{jobid}{Defined job's category ID. See glints_jobid() function.}

\item{limit}{Limit amount of job result}
}
\value{
Job vacancy data.frame in tibble format
}
\description{
Get job vacancy from Glints' website
}
\examples{
\dontrun{
glints(2, 15) # return data science job
}
}
