% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glints.R
\name{glints}
\alias{glints}
\title{Glints Vacancy}
\usage{
glints(key, limit = 30L)
}
\arguments{
\item{key}{A character indicate the keyword for job search.}

\item{limit}{Limit amount of job results to return. Maximum 100.}
}
\description{
Get job vacancy from Glints' website.
}
\examples{
\dontrun{
glints("data analyst", 15) # return data analyst job
}

}
