% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glints.R
\name{glints}
\alias{glints}
\title{Glints Vacancy}
\usage{
glints(key, limit = 30L)
}
\arguments{
\item{key}{(character) Keyword for the jobs}

\item{limit}{(numeric) Limit amount of job results}
}
\value{
(data.frame) Job vacancy data.frame in tibble format
}
\description{
Get job vacancy from Glints' website
}
\examples{
\dontrun{
glints("data analyst", 15) # return data analyst job
}

}
