% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glints.R
\name{glints}
\alias{glints}
\title{Get Vacancy from Glints}
\usage{
glints(jobid = 2, lim = 30)
}
\arguments{
\item{jobid}{Job's id}

\item{lim}{Limit number of job result}
}
\value{
Job vacancy data.frame in tibble format
}
\description{
Get Vacancy from Glints
}
\examples{
\dontrun{
glints(2, 15)
}
}
