% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gql.R
\name{gql}
\alias{gql}
\title{Request GraphQL Query}
\usage{
gql(
  query,
  ...,
  .token = NULL,
  .variables = NULL,
  .operationName = NULL,
  .url = url
)
}
\arguments{
\item{query}{GraphQL query}

\item{...}{Other arguments used by httr::POST()}

\item{.token}{Token}

\item{.variables}{Variables}

\item{.operationName}{Operation name}

\item{.url}{Complete URL to RestAPI service base to endpoint}
}
\value{
List of query result
}
\description{
Request data using GraphQL query
}
