% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indeed.R
\name{indeed}
\alias{indeed}
\title{Indeed Indonesia Vacancy}
\usage{
indeed(key, page = 2)
}
\arguments{
\item{key}{Keyword for the jobs}

\item{page}{A numeric indicate number of pages to be looking at}
}
\value{
Job opportunity data.frame in tibble format
}
\description{
Get job vacancy from Indeed's website (Indonesia)
}
\examples{
\dontrun{
indeed("data analyst") # return data analyst job
}
}
