% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobstreet.R
\name{jobstreet}
\alias{jobstreet}
\title{Jobstreet Vacancy}
\usage{
jobstreet(key, limit = 30L)
}
\arguments{
\item{key}{(character) Keyword for the jobs}

\item{limit}{(numeric) Limit amount of job results}
}
\value{
Job vacancy data.frame in tibble format
}
\description{
Get job vacancy from Jobstreet's website
}
\examples{
\dontrun{
jobstreet("data analyst") # return search result for data analyst
jobstreet("data engineer", 10) # return 10 data engineer jobs
}

}
