% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobstreet.R
\name{jobstreet}
\alias{jobstreet}
\title{Jobstreet Vacancy}
\usage{
jobstreet(key, limit = 30L)
}
\arguments{
\item{key}{Characters indicate the keyword for job search}

\item{limit}{Number of limit for job results}
}
\description{
Get job vacancy from Jobstreet's website (Seek ID).
}
\examples{
\dontrun{
jobstreet("data analyst") # return search result for data analyst
jobstreet("data engineer", 10) # return 10 data engineer jobs
}

}
