% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobstreet.R
\name{jobstreet}
\alias{jobstreet}
\title{Jobstreet Vacancy}
\usage{
jobstreet(key, page = 1)
}
\arguments{
\item{key}{Keyword for searching}

\item{page}{The amount of pages to be looking at}
}
\value{
Job opportunity data.frame in tibble format
}
\description{
Get job vacancy from Jobstreet's website
}
\examples{
\dontrun{
jobstreet("data analyst") # return search result for data analyst
}
}
