% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalibrr.R
\name{kalibrr}
\alias{kalibrr}
\title{Kalibrr Vacancy}
\usage{
kalibrr(key, limit = 30L, remote = FALSE)
}
\arguments{
\item{key}{A character indicate the keyword for job search.}

\item{limit}{Limit amount of job results to return.}

\item{remote}{Remote work allowance.}
}
\description{
Get job vacancy from Kalibbr ID website.
}
\examples{
\dontrun{
kalibrr("data engineer", 20, TRUE) # return remote data engineer job
}

}
