% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restruct_job.R
\name{restruct_job}
\alias{restruct_job}
\title{Restructuring List to Dataframe}
\usage{
restruct_job(jobls)
}
\arguments{
\item{jobls}{List of job result from GQL}
}
\value{
A tibble of job
}
\description{
Restructuring List to Dataframe
}
