% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restruct_job.R
\name{restruct_job}
\alias{restruct_job}
\title{Restructuring List to Dataframe}
\usage{
restruct_job(jobls)
}
\arguments{
\item{jobls}{A list of job result from \code{gql()} function}
}
\value{
A data.frame of job in tibble format
}
\description{
Restructuring List to Dataframe
}
